refactor: move DocumentHighlightAdapter to new line
